Convert dimensions from millimeters to inches
The dimensions are much simpler in inches, which are used in the
original English edition, than after their translation to millimeters
for the French edition.

I also configured the viewport of the image to use units 10 times finer,
to avoid the need for floating point values most of the time.