Step 9: draw stars joining points at 1.75in/2.25in to points at 3 in
The first star joins up points at 1.75 inches from the center on the
long lines to points at 3 inches from the center on the short lines.

The second star joins up points at 2.25 inches from the center on the
long lines to points at 3 inches from the center on the short lines.