Step 10: draw stars within and outside of double interior ring
The star outside the double interior ring joins up points at 0.65 inch
from the center on the long lines to points at 1.5 inches from the center
on the short lines.

The star inside the double interior ring joins up points at 0.25 inch
from the center on the long lines to points to points in the middle
of edges of the double interior ring (hexagon with sides of 0.65 inch).

Besides, a stick star is drawn within the latter star, with three lines
of 0.5 inch of length each rotated in steps of 60 degrees.